+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Change log
 
 ## [Unreleased]
-[Unreleased]: https://github.com/JakeWharton/test-distribution-gradle-plugin/compare/0.3.0...HEAD
+[Unreleased]: https://github.com/JakeWharton/test-distribution-gradle-plugin/compare/0.4.0...HEAD
+
+Nothing yet!
+
+
+## [0.4.0] - 2026-09-02
+[0.4.0]: https://github.com/JakeWharton/test-distribution-gradle-plugin/releases/tag/0.4.0
 
 Added:
 - Support for multiple JVM test runs for Kotlin multiplatform plugin.
@@ -11,7 +17,7 @@ Fixed:
 - Android distribution tasks are now compatible with configuration cache.
 
 
-## [0.3.0] - 2026-01-016
+## [0.3.0] - 2026-01-16
 [0.3.0]: https://github.com/JakeWharton/test-distribution-gradle-plugin/releases/tag/0.3.0
 
 Added: